feat(ritk-filter): ShotNoise float-exact -- completes the RNG noise family
Fourth and final RNG filter unblocked. Rewrote noise/shot.rs to ITK's
exact functor: in = scale*input; Knuth Poisson over the MT19937 stream for
in<50 (output (k-1)/scale), Normal approximation in + sqrt(in)*FastNorm()
for in>=50 (output /scale). Both generators (MersenneTwister + FastNorm)
seeded from Hash(userSeed,0), stepped only on the branch taken. seed
u64->u32; scale<=0 guard. Removed the now-unused box_muller helper (all
noise filters use exact ITK generators).

<=1e-3 vs sitk.ShotNoise single-threaded across 3 scale/seed cases (arange
image spans the lambda=50 threshold, exercising both branches). Coverage
272 -> 273/298. ALL FOUR RNG noise filters now bit/float-exact -- the
prior "non-deterministic/impossible" verdict is fully disproven.

+++ crates/ritk-python/src/filter/noise.rs
@@ -87,8 +87,8 @@ pub fn salt_and_pepper_noise(
 /// Returns:
 ///     PyImage with Poisson noise.
 #[pyfunction]
-#[pyo3(signature = (image, scale, seed=42_u64))]
-pub fn shot_noise(py: Python<'_>, image: &PyImage, scale: f64, seed: u64) -> RitkResult<PyImage> {
+#[pyo3(signature = (image, scale, seed=42_u32))]
+pub fn shot_noise(py: Python<'_>, image: &PyImage, scale: f64, seed: u32) -> RitkResult<PyImage> {
     let img = Arc::clone(&image.inner);
     let result = py
         .allow_threads(|| {
